Session security

+++ backend/submit/remove-course.php
@@ -1,10 +1,23 @@
 <?php
 $course = $_POST['course'];
-include $_SERVER['DOCUMENT_ROOT'] . '/CSE442-542/2018-Summer/team03/backend/sql/database.php';
-include $_SERVER['DOCUMENT_ROOT'] . '/CSE442-542/2018-Summer/team03/backend/sql/credentials.php';
 
-$database = new Database($CREDENTIALS["address"],$CREDENTIALS["database"], $CREDENTIALS["username"], $CREDENTIALS["password"]);
-$database->connect();
+include $_SERVER['DOCUMENT_ROOT'] . '/CSE442-542/2018-Summer/team03/backend/session/session.php';
+
+if (!$logged_in) {
+  header('Location: ../../frontend/course.html?course=' . $course);
+  exit();
+}
+
+$b = false;
+$courses = $database->getCourses($user_id);
+for($x = 0; $x < sizeof($courses); $x++)
+  if($courses[$x]['id'] == $course)
+    $b = true;
+if(!$b) {
+  header('Location: ../../frontend/course.html?course=' . $course);
+  exit();
+}
+
 $database->removeCourse($course);
 header("Location: ../../frontend/index.html");
 exit();
